+++ src/classes/CoreCache.ts
@@ -7,10 +7,10 @@ export default abstract class CoreCache
   implements ICoreCache
 {
   constructor(
-    chanel: string,
+    channel: string,
     module: ICoreKernelModule<any, any, any, any, any>
   ) {
-    super(`client-${chanel}`, module);
+    super(`client-${channel}`, module);
   }
 
   abstract start(): Promise<void>;