+++ src/utils/index.ts
@@ -1,6 +1,6 @@
 import initHandler from './initHandler';
-import { getTimeStamp, timePrint } from './timeUtil';
 
+export * from './timeUtil';
 export * from './envUtil';
 export * from './fsUtil';
-export { initHandler, timePrint, getTimeStamp };
+export { initHandler };